usual gcc fixes

+++ src/NoLifeStory/Time.cpp
@@ -24,5 +24,5 @@ void NLS::_Time::Step() {
 		C("INFO") << "Time taken: " << delta << " ms" << endl;
 		output = false;
 	}
-	delta = min(100, delta);
+	delta = min((uint32_t)100, delta);
 }
